feat(layout): refactor layout structure and introduce ClientLayout component
- Removed unnecessary client-side logic from RootLayout in layout.tsx.
- Introduced a new ClientLayout component to handle user authentication and navigation.
- Integrated user data fetching and session management into ClientLayout.
- Simplified the RootLayout to focus on rendering the main structure and analytics.

These changes enhance the modularity of the layout by separating concerns, improving maintainability, and streamlining the user authentication process. The introduction of the ClientLayout component allows for better organization of user-related functionalities, leading to a cleaner and more efficient codebase.
